Rework a bit openapi package structure
